% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_fun.R
\name{box_fun}
\alias{box_fun}
\title{box_fun}
\usage{
box_fun(d, i, annotate)
}
\arguments{
\item{d}{data frame}

\item{i}{column index}

\item{annotate}{optional data frame with annotation information in the following format (order is important): Variable, Norm_min, Norm_max, Min, Max}
}
\value{
png image(s)
}
\description{
Create boxplots and choose number to display per page
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2, gridExtra
}
\examples{
box_fun(d, i, annotate)
}
