% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{box_plot}
\alias{box_plot}
\title{box_plot}
\usage{
box_plot(d, n = 12, file = "plot", nrow = 4, ncol = 3, wi = 13.5,
  hgt = 12, res = 300, annotate)
}
\arguments{
\item{d}{data frame}

\item{n}{number of plots to display per page}

\item{nrow}{number of rows to display per page}

\item{ncol}{number of columns to display per page}

\item{wi}{width of plot}

\item{hgt}{height of plot}

\item{res}{resolution of plot}

\item{annotate}{optional data frame with clinical lab information}
}
\value{
png image(s)
}
\description{
Create boxplots and choose number to display per page
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2, gridExtra
}
\examples{
box_plot(d, n=12, file="plot", nrow=4, ncol=3, wi=13.5, hgt=12, res=210, annotate)
}
