% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_tests.R
\name{chisq_tests}
\alias{chisq_tests}
\title{chisq_tests}
\usage{
chisq_tests(df)
}
\arguments{
\item{df}{data frame, with IID as first column}
}
\value{
data frame with two columns, Variable and P.Value
}
\description{
Create table of chi-square tests for each variable
}
\examples{
chisq_tests(df)
}
