% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colfilter.R
\name{colfilter}
\alias{colfilter}
\title{colfilter}
\usage{
colfilter(d, cols, exclude = FALSE)
}
\arguments{
\item{d}{data frame}

\item{cols}{list of variable names to subset}

\item{exclude}{boolean, keep or exclude}
}
\value{
dataframe
}
\description{
Subset data by column name
}
\examples{
colfilter(d, cols, exclude=FALSE)
}
