% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlations}
\alias{correlations}
\title{correlations}
\usage{
correlations(df, x = 0.75)
}
\arguments{
\item{df}{data frame}

\item{x}{correlation threshold, -1 <= x <= 1}
}
\value{
data frame with three columns, Variable 1, Variable 2, and the correlation between Variable 1 and Variable 2
}
\description{
Identify variables that are correlated above a certain threshold
Note: Uses Pearson/pairwise.complete.obs in the base R cor function
Note: A threshold of 0.75 also includes variables with a negative correlation stronger than -0.5
}
\examples{
require(NHANES)
data(NHANES)
correlations(get_continuous(NHANES))
}
\seealso{
Other summary functions: \code{\link{chisq_tests}},
  \code{\link{freq_tables}}, \code{\link{get_uniq}},
  \code{\link{outlier_impact}}, \code{\link{sample_size}}

Other continuous variable functions: \code{\link{box_plot}},
  \code{\link{hist_fun}}, \code{\link{multi_plot}},
  \code{\link{outlier_impact}}, \code{\link{qq_fun}},
  \code{\link{qq_plot}}, \code{\link{remove_outliers}}
}
