% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eman.R
\name{eman}
\alias{eman}
\title{eman}
\usage{
eman(d, ewas = TRUE, groups = NULL, line = NULL, title = NULL,
  morecolors = FALSE, file = "eman", hgt = 7, wi = 12, res = 300)
}
\arguments{
\item{d}{data frame, if ewas=FALSE columns one and two must be Variable and pvalue, Shape is optional 3rd column}

\item{ewas}{boolean if d frame is in ewas output format, default TRUE}

\item{groups}{optional file containing variable grouping information}

\item{line}{optional pvalue threshold to draw red line at}

\item{title}{optional string for plot title}

\item{morecolors}{boolean, expand color palette, requires RColorBrewer package, default FALSE}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}
}
\value{
png image(s)
}
\description{
Create Manhattan plots for EWAS\cr
Note: There is an issue with dev.off() if using RStudio\cr
Dependencies: ggplot2\cr
Suggested: RColorBrewer
}
\examples{
\dontrun{
eman(d, ewas=TRUE, groups=NULL, line, title=NULL, morecolors=FALSE,
file="eman", hgt=7, wi=12, res=300 )
}
}
\seealso{
Other plot functions: \code{\link{bar_plot}},
  \code{\link{box_plot}}, \code{\link{hist_fun}},
  \code{\link{hist_plot}}, \code{\link{multi_plot}},
  \code{\link{qq_fun}}, \code{\link{qq_plot}}
}
\concept{plot functions}
