% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_tables.R
\name{freq_tables}
\alias{freq_tables}
\title{freq_tables}
\usage{
freq_tables(df)
}
\arguments{
\item{df}{data frame with ID as first column}
}
\value{
data frame with three columns, Variable, Category, and N
}
\description{
Create frequency tables for each variable\cr
Gives the number of observations in each category for all variables in file
}
\examples{
require(NHANES)
data(NHANES)
freq_tables(get_categorical(NHANES))
}
\seealso{
Other summary functions: \code{\link{chisq_tests}},
  \code{\link{correlations}}, \code{\link{get_uniq}},
  \code{\link{outlier_impact}}, \code{\link{sample_size}}

Other categorical variable functions: \code{\link{bar_plot}},
  \code{\link{chisq_tests}}
}
