% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binary.R
\name{get_binary}
\alias{get_binary}
\title{get_binary}
\usage{
get_binary(df, lev = 2)
}
\arguments{
\item{df}{data frame of mixed variable types}

\item{lev}{number of levels desired}
}
\value{
data frame containing only IID and variables with the number of levels specified
}
\description{
Identify binary variables from a data frame of mixed variable types
Comparison operator used is =
NA values do not contribute to level counts
}
\examples{
get_binary(df, lev=2)
}
