% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_check.R
\name{get_check}
\alias{get_check}
\title{get_check}
\usage{
get_check(df, lower = 6, upper = 18)
}
\arguments{
\item{df}{data frame of mixed variable types}

\item{lev}{number of levels desired}
}
\value{
data frame containing only IID and variables meeting criteria, which should then be assessed manually
}
\description{
Identify ambiguous type variables from a data frame of mixed variable types
Comparison operators used are < and >
NA values do not contribute to level counts
}
\examples{
get_check(df, lower=6, upper=18)
}
