% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_continuous.R
\name{get_continuous}
\alias{get_continuous}
\title{get_continuous}
\usage{
get_continuous(df, lower = 15)
}
\arguments{
\item{df}{data frame of mixed variable types}

\item{lower}{minimum number of unique values to be considered continuous}
}
\value{
data frame containing only ID and variables meeting the criteria
}
\description{
Identify continuous variables from a data frame of mixed variable types
Comparison operators used are >=
NA values do not contribute to level counts
}
\examples{
get_continuous(df, lower=15)
}
