% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_levels.R
\name{get_levels}
\alias{get_levels}
\title{get_levels}
\usage{
get_levels(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame with two columns, one for variable name and one for number of unique values
}
\description{
Get number of unique values in each column, not including 'NA'
}
\examples{
get_levels(df)
}
