% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uniq.R
\name{get_uniq}
\alias{get_uniq}
\title{get_uniq}
\usage{
get_uniq(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame with two columns, one for variable name and one for number of unique values
}
\description{
Get number of unique values in each column, not including 'NA'
}
\examples{
require(NHANES)
data(NHANES)
get_uniq(NHANES)
}
\seealso{
Other summary functions: \code{\link{chisq_tests}},
  \code{\link{correlations}}, \code{\link{freq_tables}},
  \code{\link{outlier_impact}}, \code{\link{sample_size}}
}
\concept{summary functions}
