% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_fun.R
\name{hist_fun}
\alias{hist_fun}
\title{hist_fun}
\usage{
hist_fun(d, i, annotate)
}
\arguments{
\item{d}{data frame}

\item{i}{column name}

\item{annotate}{optional data frame of annotation information in the following format (order is important): Variable, Norm_min, Norm_max, Min, Max}
}
\value{
ggplot object
}
\description{
Create histograms and choose number to display per page
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2, gridExtra
}
\examples{
hist_fun(d, i, annotate)
}
