% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_n.R
\name{min_n}
\alias{min_n}
\title{min_n}
\usage{
min_n(df, n = 200)
}
\arguments{
\item{df}{data frame}

\item{n}{minimum sample size desired}
}
\value{
data frame containing only those variables with at least n samples
}
\description{
Keep variables with a minimum of n samples
}
\examples{
min_n(df, n)
}
