% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_plot.R
\name{multi_plot}
\alias{multi_plot}
\title{multi_plot}
\usage{
multi_plot(d, n = 6, file = "plot", nrow = 3, ncol = 2, wi = 13.5,
  hgt = 9, res = 210, type, annotate)
}
\arguments{
\item{d}{data frame}

\item{n}{number of variables to display per page}

\item{nrow}{number of variables per row}

\item{ncol}{number of variables per column}

\item{wi}{width of plot}

\item{hgt}{height of plot}

\item{res}{resolution of plot}

\item{type}{sepcify plot type 'hist-qq, box-qq, hist-box, hist-box-qq'}

\item{annotate}{optional data frame with clinical lab information}
}
\value{
png image(s)
}
\description{
Create multiple types of plots on the same page
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2, gridExtra
}
\examples{
multi_plot(d, n=6, file="plot", nrow=3, ncol=2, wi=13.5, hgt=9, res=210, type="hist-qq", annotate)
}
