% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_threshold.R
\name{outlier_threshold}
\alias{outlier_threshold}
\title{outlier_threshold}
\usage{
outlier_threshold(df, x = 2.5)
}
\arguments{
\item{df}{data frame}

\item{x}{number of standard deviations away from the mean to qualify as an outlier}
}
\value{
data frame consisting of summary statistics for the input dataframe and potential dataframe if outliers were to be removed at that threshold
}
\description{
Create data frame with outliers and summary statistics for each variable
}
\examples{
outlier_threshold(df, x=2.5)
}
