% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_fun.R
\name{qq_fun}
\alias{qq_fun}
\title{qq_fun}
\usage{
qq_fun(d, i)
}
\arguments{
\item{d}{data frame}

\item{i}{column name}
}
\value{
ggplot object
}
\description{
Create qqplots and choose number to display per page\cr
Note: Call quatrz() prior to plottting in RStudio\cr
Dependencies: ggplot2, gridExtra
}
\seealso{
Other plot functions: \code{\link{bar_plot}},
  \code{\link{box_plot}}, \code{\link{eman}},
  \code{\link{hist_fun}}, \code{\link{hist_plot}},
  \code{\link{multi_plot}}, \code{\link{qq_plot}}

Other continuous variable functions: \code{\link{box_plot}},
  \code{\link{correlations}}, \code{\link{hist_fun}},
  \code{\link{multi_plot}}, \code{\link{outlier_impact}},
  \code{\link{qq_plot}}, \code{\link{remove_outliers}}
}
\concept{continuous variable functions}
\concept{plot functions}
