% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{qq_plot}
\alias{qq_plot}
\title{qq_plot}
\usage{
qq_plot(d, n = 12, file = "plot", nrow = 4, ncol = 3, wi = 13.5,
  hgt = 12, res = 300)
}
\arguments{
\item{d}{data frame}

\item{n}{number of plots to display per page}

\item{file}{filename of output}

\item{nrow}{number of rows to display per page}

\item{ncol}{number of columns to display per page}

\item{wi}{width of plot}

\item{hgt}{height of plot}

\item{res}{resolution of plot}
}
\value{
png image(s)
}
\description{
Create qqplots and choose number to display per page\cr
Note: Call quartz() prior to plotting if using RStudio\cr
Dependencies: ggplot2, gridExtra
}
\examples{
\dontrun{
qq_plot(d, n=12, file="plot", nrow=4, ncol=3, wi=13.5, hgt=12, res=210)
}
}
\seealso{
Other plot functions: \code{\link{bar_plot}},
  \code{\link{box_plot}}, \code{\link{eman}},
  \code{\link{hist_fun}}, \code{\link{hist_plot}},
  \code{\link{multi_plot}}, \code{\link{qq_fun}}

Other continuous variable functions: \code{\link{box_plot}},
  \code{\link{correlations}}, \code{\link{hist_fun}},
  \code{\link{multi_plot}}, \code{\link{outlier_impact}},
  \code{\link{qq_fun}}, \code{\link{remove_outliers}}
}
\concept{continuous variable functions}
\concept{plot functions}
