% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_key.R
\name{recode_key}
\alias{recode_key}
\title{recode_key}
\usage{
recode_key(df, key)
}
\arguments{
\item{df}{data frame}

\item{key}{data frame with two columns, "Variable" and "Missing Value"}
}
\value{
data frame with all missing values recoded to NA
}
\description{
Recode based on a data key if multiple columns have different missing values
Note: this is not recommended if the value to replace is a float
}
\examples{
recode_key(df, key)
}
