% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_missing.R
\name{recode_missing}
\alias{recode_missing}
\title{recode_missing}
\usage{
recode_missing(df, na_val)
}
\arguments{
\item{df}{data frame}

\item{na_val}{value to replace with NA}
}
\value{
data frame with missing value recoded to NA
}
\description{
Recode if there is one missing value for entire dataset
Note: this is not recommended if the value to be changed to NA is a float
}
\examples{
recode_missing(df, na_val)
}
