% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{remove_outliers}
\usage{
remove_outliers(df, x = 2.5)
}
\arguments{
\item{df}{data frame with first column as ID}

\item{x}{number of standard deviations away from the mean to qualify as an outlier}
}
\value{
data frame with outlier values within each column replaced with NA
}
\description{
Remove outliers from data frame
This function does not remove the entire row of data, only replaces the outlier value with NA in a particular column
}
\examples{
remove_outliers(df, x=2.5)
}
