% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowfilter.R
\name{rowfilter}
\alias{rowfilter}
\title{rowfilter}
\usage{
rowfilter(d, samples, exclude = FALSE)
}
\arguments{
\item{d}{data frame, with first column as ID}

\item{samples}{list or dataframe of samples to subset}

\item{exclude}{boolean, keep or exclude}
}
\value{
dataframe
}
\description{
Subset data by sample name
}
\examples{
rowfilter(d, samples, exclude=FALSE)
}
