% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_keep.R
\name{sample_keep}
\alias{sample_keep}
\title{sample_keep}
\usage{
sample_keep(df, n = 200)
}
\arguments{
\item{df}{data frame}

\item{n}{minimum sample size desired}
}
\value{
data frame containing only those variables with at least n samples
}
\description{
Keep variables with a minimum of n samples
}
\examples{
sample_keep(df, n)
}
