% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_keep_subgroups.R
\name{sample_keep_subgroups}
\alias{sample_keep_subgroups}
\title{sample_keep_subgroups}
\usage{
sample_keep_subgroups(df, n = 200)
}
\arguments{
\item{df}{data frame with IID as first column}

\item{n}{minimum sample size desired}
}
\value{
data frame containing only those variables with at least n samples per factor level
}
\description{
Keep variables with a minimum of n samples
}
\examples{
sample_keep_subgroups(df, n)
}
