% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplefilter.R
\name{samplefilter}
\alias{samplefilter}
\title{samplefilter}
\usage{
samplefilter(d, samples, exclude = FALSE)
}
\arguments{
\item{d}{data frame, with first column as IID}

\item{samples}{list or dataframe of samples to subset}

\item{exclude}{boolean, keep or exclude}
}
\value{
dataframe
}
\description{
Subset data by sample name
}
\examples{
samplefilter(d, samples, exclude=FALSE)
}
