% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_list.R
\name{transform_list}
\alias{transform_list}
\title{transform_list}
\usage{
transform_list(df, key)
}
\arguments{
\item{df}{data frame}

\item{key}{data frame with columns for variable and transform type}
}
\value{
data frame with the variables transformed according to the key
}
\description{
Transform user list of columns and transform type for each column (sqrt,log)
Can mix transform types but can only do one transform per variable at a time
}
\examples{
transform_list(df, key)
}
