% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transvar.R
\name{transvar}
\alias{transvar}
\title{transvar}
\usage{
transvar(df, key)
}
\arguments{
\item{df}{data frame}

\item{key}{data frame with columns for variable and transform type}
}
\value{
data frame with the variables transformed according to the key
}
\description{
Transform user list of columns and transform type for each column (sqrt,log)
Can mix transform types but can only do one transform per variable at a time
}
\examples{
transvar(df, key)
}
