% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{box_plot}
\alias{box_plot}
\title{box_plot}
\usage{
box_plot(d, n = 12, file = "plot", nrow = 4, ncol = 3, wi = 13.5,
  hgt = 12, res = 300, annotate)
}
\arguments{
\item{d}{data frame}

\item{n}{number of plots to display per page}

\item{file}{filename of output}

\item{nrow}{number of rows to display per page}

\item{ncol}{number of columns to display per page}

\item{wi}{width of plot}

\item{hgt}{height of plot}

\item{res}{resolution of plot}

\item{annotate}{optional data frame with clinical lab information}
}
\value{
png image(s)
}
\description{
Create boxplots and choose number to display per page\cr
Note: Call quatrz() prior to plottting in RStudio\cr
Dependencies: ggplot2, gridExtra
}
\examples{
\dontrun{
box_plot(d, n=12, file="plot", nrow=4, ncol=3, wi=13.5, hgt=12, res=210, annotate)
}
}
\seealso{
Other plot functions: \code{\link{bar_plot}},
  \code{\link{eman}}, \code{\link{hist_fun}},
  \code{\link{hist_plot}}, \code{\link{multi_plot}},
  \code{\link{qq_fun}}, \code{\link{qq_plot}}

Other continuous variable functions: \code{\link{correlations}},
  \code{\link{hist_fun}}, \code{\link{multi_plot}},
  \code{\link{outlier_impact}}, \code{\link{qq_fun}},
  \code{\link{qq_plot}}, \code{\link{remove_outliers}}
}
