% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_tests.R
\name{chisq_tests}
\alias{chisq_tests}
\title{chisq_tests}
\usage{
chisq_tests(df)
}
\arguments{
\item{df}{data frame, with ID as first column\cr}
}
\value{
data frame with two columns, Variable and P.Value
}
\description{
Create table of chi-square tests for each variable\cr
}
\examples{
\dontrun{
chisq_tests(df)
}
}
\seealso{
Other summary functions: \code{\link{correlations}},
  \code{\link{freq_tables}}, \code{\link{get_uniq}},
  \code{\link{outlier_impact}}, \code{\link{sample_size}}

Other categorical variable functions: \code{\link{bar_plot}},
  \code{\link{freq_tables}}
}
