% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colfilter.R
\name{colfilter}
\alias{colfilter}
\title{colfilter}
\usage{
colfilter(d, cols, exclude = FALSE)
}
\arguments{
\item{d}{data frame}

\item{cols}{list or dataframe of variable names to subset}

\item{exclude}{boolean, keep or exclude}
}
\value{
dataframe
}
\description{
Subset data by column name
}
\examples{
require(NHANES)
data(NHANES)
cols <- c("SurveyYr","Gender", "Age")
head(colfilter(NHANES, cols=cols))
}
\seealso{
Other filter functions: \code{\link{get_binary}},
  \code{\link{get_categorical}}, \code{\link{get_check}},
  \code{\link{get_continuous}}, \code{\link{min_cat_n}},
  \code{\link{min_n}}, \code{\link{remove_outliers}},
  \code{\link{rowfilter}}
}
\concept{filter functions}
