% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlations}
\alias{correlations}
\title{correlations}
\usage{
correlations(df, x = 0.5)
}
\arguments{
\item{df}{data frame}

\item{x}{correlation threshold, -1 <= x <= 1}
}
\value{
data frame with three columns, Variable 1, Variable 2, and the correlation between Variable 1 and Variable 2
}
\description{
Identify variables that are correlated above a certain threshold
Note: Uses the pairwise.complete.obs in the base R cor function
Note: A threshold of 0.5 also includes variables with a negative correlation stronger than -0.5
}
\examples{
correlations(df, x=0)
}
