% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas.R
\name{ewas}
\alias{ewas}
\title{ewas}
\usage{
ewas(d, cat_vars = NULL, cont_vars = NULL, y, cov = NULL, regress)
}
\arguments{
\item{d}{data.frame or survey::svydesign object containing all of the data}

\item{cat_vars}{List of variables to regress that are categorical or binary}

\item{cont_vars}{List of variables to regress that are continuous}

\item{y}{name(s) of response variable(s)}

\item{cov}{vector containing names of covariates}

\item{regress}{family for the regression model as specified in glm, linear or logisitic}
}
\value{
data frame containing following fields Variable, Sample Size, Converged, SE, Beta, Variable p-value, LRT, AIC, pval, Phenotype
}
\description{
Run environment-wide association study
}
\examples{
\dontrun{
ewas(d, cat_vars, cont_vars, y, cov, regress)
}
}
\seealso{
Other analysis functions: \code{\link{ewas_pval_adjust}}
}
\concept{analysis functions}
