% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas.R
\name{ewas}
\alias{ewas}
\title{ewas}
\usage{
ewas(d, cat_vars = NULL, cont_vars = NULL, y, cat_covars = NULL,
  cont_covars = NULL, regress = "gaussian",
  allowed_nonvarying = NULL)
}
\arguments{
\item{d}{data.frame or survey::svydesign object containing all of the data}

\item{cat_vars}{List of variables to regress that are categorical or binary}

\item{cont_vars}{List of variables to regress that are continuous}

\item{y}{name(s) of response variable(s)}

\item{cat_covars}{List of covariates that are categorical or binary}

\item{cont_covars}{List of covariates that are continuous}

\item{regress}{family for the regression model as specified in glm ('gaussian' by default)}

\item{allowed_nonvarying}{list of covariates that are excluded from the regression when they do not vary instead of returning a NULL result.}
}
\value{
data frame containing following fields Variable, Sample Size, Converged, SE, Beta, Variable p-value, LRT, AIC, pval, phenotype
}
\description{
Run environment-wide association study
}
\examples{
\dontrun{
ewas(d, cat_vars, cont_vars, y, cat_covars, cont_covars, regress)
}
}
\seealso{
Other analysis functions: \code{\link{ewas_pval_adjust}},
  \code{\link{weighted_ewas}}
}
\concept{analysis functions}
