% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas.R
\name{ewas}
\alias{ewas}
\title{ewas}
\usage{
ewas(cat = NULL, cont = NULL, y, cov = NULL, regress, adjust)
}
\arguments{
\item{cat}{data frame containing categorical variables with first column as IID}

\item{cont}{data frame containing continuous variables with first column as IID}

\item{y}{name of response variable}

\item{cov}{vector containing names of covariates}

\item{regress}{family for the regression model as specified in glm, linear or logisitic}

\item{adjust}{p-value adjustment (bonferroni or fdr)}
}
\value{
data frame containing following fields Variable, Variable Type, Regression Type, Sample Size, SE, Beta, Variable p-value, LRT, AIC, adjusted p-value
}
\description{
Run environment-wide association study
}
\examples{
ewas(cat, cont, y, cov, regress, adjust)
}
