% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas.R
\name{ewas}
\alias{ewas}
\title{ewas}
\usage{
ewas(d, cat_vars = NULL, cont_vars = NULL, y, cat_covars = NULL,
  cont_covars = NULL, regression_family = "gaussian",
  allowed_nonvarying = NULL, min_n = 200, weights = NULL,
  ids = NULL, strata = NULL, fpc = NULL, ...)
}
\arguments{
\item{d}{data.frame containing all of the data}

\item{cat_vars}{List of variables to regress that are categorical or binary}

\item{cont_vars}{List of variables to regress that are continuous}

\item{y}{name(s) of response variable(s)}

\item{cat_covars}{List of covariates that are categorical or binary}

\item{cont_covars}{List of covariates that are continuous}

\item{regression_family}{family for the regression model as specified in glm ('gaussian' by default)}

\item{allowed_nonvarying}{list of covariates that are excluded from the regression when they do not vary instead of returning a NULL result.}

\item{min_n}{minimum number of observations required (after dropping those with NA values) before running the regression (200 by default)}

\item{weights}{NULL by default (for unweighted).  May be set to a string name of a single weight to use for every variable, or a named list that maps variable names to the weights that should be used for that variable's regression}

\item{ids}{NULL by default (for no clusters).  May be set to a string name of a column in the data which provides cluster IDs.}

\item{strata}{NULL by default (for no strata).  May be set to a string name of a column in the data which provides strata IDs.}

\item{fpc}{NULL by default (for no fpc).  May be set to a string name of a column in the data which provides fpc values.}

\item{...}{other arguments passed to svydesign which are ignored if 'weights' is NULL}
}
\value{
data frame containing following fields Variable, Sample Size, Converged, SE, Beta, Variable p-value, LRT, AIC, pval, phenotype, weight
}
\description{
Run environment-wide association study, optionally using \code{\link[survey]{svydesign}} from the \pkg{survey} package
Note: It is possible to specify \emph{ids} and/or \emph{strata}.  When \emph{ids} is specified without \emph{strata}, 
the standard error is infinite and the anova calculation for categorical variables fails.  This is due to the 
\href{http://r-survey.r-forge.r-project.org/survey/exmample-lonely.html}{lonely psu} problem.
}
\examples{
\dontrun{
ewas(d, cat_vars, cont_vars, y, cat_covars, cont_covars, regression_family)
}
}
\seealso{
Other analysis functions: \code{\link{ewas_pval_adjust}}
}
\concept{analysis functions}
