% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_tables.R
\name{freq_tables}
\alias{freq_tables}
\title{freq_tables}
\usage{
freq_tables(df)
}
\arguments{
\item{df}{data frame with ID as first column}
}
\value{
data frame with three columns, Variable, Category, and N
}
\description{
Create frequency tables for each variable
Gives the number of observations in each category for all variables in file
}
\examples{
freq_tables(df)
}
