% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binary.R
\name{get_binary}
\alias{get_binary}
\title{get_binary}
\usage{
get_binary(df, lev = 2)
}
\arguments{
\item{df}{data frame of mixed variable types}

\item{lev}{number of levels desired}
}
\value{
data frame containing only ID and variables with the number of levels specified
}
\description{
Identify binary variables from a data frame of mixed variable types
Comparison operator used is =
NA values do not contribute to level counts
}
\examples{
require(NHANES)
data(NHANES)
head(get_binary(NHANES))
}
\seealso{
Other filter functions: \code{\link{colfilter}},
  \code{\link{get_categorical}}, \code{\link{get_check}},
  \code{\link{get_continuous}}, \code{\link{min_cat_n}},
  \code{\link{min_n}}, \code{\link{remove_outliers}},
  \code{\link{rowfilter}}
}
