% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categorical.R
\name{get_categorical}
\alias{get_categorical}
\title{get_categorical}
\usage{
get_categorical(df, lower = 3, upper = 6)
}
\arguments{
\item{df}{data frame of mixed variable types}

\item{lower}{minimum number of unique values to be considered categorical}

\item{upper}{maximum number of unique values to be considered categorical}
}
\value{
data frame containing only ID and variables meeting the criteria
}
\description{
Identify categorical variables from a data frame of mixed variable types\cr
Comparison operators used are <= and >=\cr
NA values do not contribute to level counts
}
\examples{
require(NHANES)
data(NHANES)
head(get_categorical(NHANES))
}
\seealso{
Other filter functions: \code{\link{colfilter}},
  \code{\link{get_binary}}, \code{\link{get_check}},
  \code{\link{get_continuous}}, \code{\link{min_cat_n}},
  \code{\link{min_n}}, \code{\link{remove_outliers}},
  \code{\link{rowfilter}}
}
