% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_check.R
\name{get_check}
\alias{get_check}
\title{get_check}
\usage{
get_check(df, lower = 6, upper = 15)
}
\arguments{
\item{df}{data frame of mixed variable types}

\item{lev}{number of levels desired}
}
\value{
data frame containing only ID and variables meeting criteria, which should then be assessed manually
}
\description{
Identify ambiguous type variables from a data frame of mixed variable types
Comparison operators used are < and >
NA values do not contribute to level counts
}
\examples{
require(NHANES)
data(NHANES) 
head(get_check(NHANES))
}
\seealso{
Other filter functions: \code{\link{colfilter}},
  \code{\link{get_binary}}, \code{\link{get_categorical}},
  \code{\link{get_continuous}}, \code{\link{min_cat_n}},
  \code{\link{min_n}}, \code{\link{remove_outliers}},
  \code{\link{rowfilter}}
}
