% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_fun.R
\name{hist_fun}
\alias{hist_fun}
\title{hist_fun}
\usage{
hist_fun(d, i, annotate)
}
\arguments{
\item{d}{data frame}

\item{i}{column name}

\item{annotate}{optional data frame of annotation information in the following format (order is important): Variable, Norm_min, Norm_max, Min, Max}
}
\value{
ggplot object
}
\description{
Create histograms and choose number to display per page
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2, gridExtra
}
\seealso{
Other plot functions: \code{\link{bar_plot}},
  \code{\link{box_plot}}, \code{\link{eman}},
  \code{\link{hist_plot}}, \code{\link{multi_plot}},
  \code{\link{qq_fun}}, \code{\link{qq_plot}}

Other continuous variable functions: \code{\link{box_plot}},
  \code{\link{correlations}}, \code{\link{multi_plot}},
  \code{\link{outlier_impact}}, \code{\link{qq_fun}},
  \code{\link{qq_plot}}, \code{\link{remove_outliers}}
}
