% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_plot.R
\name{hist_plot}
\alias{hist_plot}
\title{hist_plot}
\usage{
hist_plot(d, n = 12, file = "plot", nrow = 4, ncol = 3, wi = 13.5,
  hgt = 12, res = 210, annotate)
}
\arguments{
\item{d}{data frame with variables only}

\item{n}{number of plots to display per page}

\item{nrow}{number of rows to display per page}

\item{ncol}{number of columns to display per page}

\item{wi}{width of plot}

\item{hgt}{height of plot}

\item{res}{resolution of plot}

\item{annotate}{optional data frame containing annotation information}
}
\value{
png image(s)
}
\description{
Create histograms for each column of a data frame
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2, gridExtra
}
\examples{
hist_plot(d, n=12, file="plot", nrow=4, ncol=3, wi=13.5, hgt=12, res=210, annotate)
}
