% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_plot.R
\name{multi_plot}
\alias{multi_plot}
\title{multi_plot}
\usage{
multi_plot(d, n = 6, file = "plot", nrow = 3, ncol = 2, wi = 13.5,
  hgt = 9, res = 210, type, annotate)
}
\arguments{
\item{d}{data frame}

\item{n}{number of variables to display per page}

\item{file}{filename of output}

\item{nrow}{number of variables per row}

\item{ncol}{number of variables per column}

\item{wi}{width of plot}

\item{hgt}{height of plot}

\item{res}{resolution of plot}

\item{type}{sepcify plot type 'hist-qq, box-qq, hist-box, hist-box-qq'}

\item{annotate}{optional data frame with clinical lab information}
}
\value{
png image(s)
}
\description{
Create multiple types of plots on the same page\cr
Note: Call quatrz() prior to plottting in RStudio\cr
Dependencies: gridExtra
}
\examples{
\dontrun{
multi_plot(d, n=6, file="plot", nrow=3, ncol=2, wi=13.5, hgt=9, res=210, type="hist-qq", annotate)
}
}
\seealso{
Other plot functions: \code{\link{bar_plot}},
  \code{\link{box_plot}}, \code{\link{eman}},
  \code{\link{hist_fun}}, \code{\link{hist_plot}},
  \code{\link{qq_fun}}, \code{\link{qq_plot}}

Other continuous variable functions: \code{\link{box_plot}},
  \code{\link{correlations}}, \code{\link{hist_fun}},
  \code{\link{outlier_impact}}, \code{\link{qq_fun}},
  \code{\link{qq_plot}}, \code{\link{remove_outliers}}
}
