% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_fun.R
\name{qq_fun}
\alias{qq_fun}
\title{qq_fun}
\usage{
qq_fun(d, i)
}
\arguments{
\item{d}{data frame}

\item{i}{column name}

\item{annotate}{optional data frame of annotation information}
}
\value{
ggplot object
}
\description{
Create qqplots and choose number to display per page
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2, gridExtra
}
\examples{
qq_fun(d, i, annotate)
}
