% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_missing.R
\name{recode_missing}
\alias{recode_missing}
\title{recode_missing}
\usage{
recode_missing(df, na_val)
}
\arguments{
\item{df}{data frame}

\item{na_val}{value to replace with NA}
}
\value{
data frame with missing value recoded to NA
}
\description{
Recode if there is one missing value for entire dataset\cr
Empty strings ("") and "NA" will be replaced by NA regardless of keys\cr
\strong{Note}: Replaced values must be a factor, integer, or numeric (other values are ignored).  Float values may fail to convert due to rounding.
}
\examples{
# Create an example dataset
df <- data.frame("ID" = 1:3,
                 "Age" = c(21,15, 34),
                 "Name" = c("John","Jane", "Jill"),
                 "Value1" = c("a", "b", "NA"),
                 "Value2" = c(1, 2, 99))
# Recode 99 and "NA" to NA
recoded <- recode_missing(df, 99)
}
\seealso{
Other functions for recoding missing values: \code{\link{recode_key}}
}
\concept{functions for recoding missing values}
