% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowfilter.R
\name{rowfilter}
\alias{rowfilter}
\title{rowfilter}
\usage{
rowfilter(d, samples, exclude = FALSE)
}
\arguments{
\item{d}{data frame, with first column as ID}

\item{samples}{list or dataframe of samples to subset}

\item{exclude}{boolean, keep or exclude}
}
\value{
dataframe
}
\description{
Subset data by sample name
}
\examples{
require(NHANES)
data(NHANES)
ids <- c("51624", "51625")
head(rowfilter(NHANES, samples = ids))
}
\seealso{
Other filter functions: \code{\link{colfilter}},
  \code{\link{get_binary}}, \code{\link{get_categorical}},
  \code{\link{get_check}}, \code{\link{get_continuous}},
  \code{\link{min_cat_n}}, \code{\link{min_n}},
  \code{\link{remove_outliers}}
}
\concept{filter functions}
