% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{sample_size}
\alias{sample_size}
\title{sample_size}
\usage{
sample_size(df)
}
\arguments{
\item{df}{data frame with ID as first column}
}
\value{
data frame with two columns, "Variable" and "Sample_Size"
}
\description{
Create list of sample_sizes for each variable
}
\examples{
require(NHANES)
data(NHANES)
sample_size(NHANES)
}
\seealso{
Other summary functions: \code{\link{chisq_tests}},
  \code{\link{correlations}}, \code{\link{freq_tables}},
  \code{\link{get_uniq}}, \code{\link{outlier_impact}}
}
